pyproject.toml
[project]
name = "sample-dind-lambda"
version = "0.1.0"
description = "Edge Serverless Simulator with Docker in Docker"
requires-python = ">=3.11"
dependencies = [
    # API Gateway互換サーバー
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    
    # Docker操作
    "docker>=6.1.3",
    
    # スケジューラ
    "apscheduler>=3.10.4",
    
    # JWT認証
    "pyjwt>=2.8.0",
    
    # データベース (ScyllaDB / Cassandra互換)
    # cassandra-driverはGatewayでは未使用、ビルド短縮のため削除
    
    # AWS SDK (S3/DynamoDB互換)
    "boto3>=1.34.0",
    
    # 設定管理
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "httpx>=0.25.1",
    "pyyaml>=6.0",
    
    # ロギング
    "python-json-logger>=2.0.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "respx>=0.20.2",
    "black>=23.11.0",
    "ruff>=0.1.6",
    "lefthook",  # Git hooks manager
    "cryptography>=41.0.0",  # SSL証明書生成用
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gateway"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests", "services/gateway/tests", "services/manager/tests"]
filterwarnings = [
    "ignore::urllib3.exceptions.InsecureRequestWarning",
]

